backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "realtime-chat"
version = "2.0.0"
description = "Production-ready real-time chat — FastAPI + Socket.IO + Redis"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-socketio>=5.11.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",

    "redis>=5.0.4",
    "aioredis>=2.0.1",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "alembic>=1.13.1",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
    "pytest-cov>=5.0.0",
    "python-socketio[asyncio_client]>=5.11.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
strict = true
python_version = "3.11"
